+++ src/main/java/fll/web/report/SubjectiveByJudge.java
@@ -39,6 +39,7 @@ import fll.util.FLLInternalException;
 import fll.util.FLLRuntimeException;
 import fll.util.FOPUtils;
 import fll.util.FP;
+import fll.util.FOPUtils.Margins;
 import fll.web.ApplicationAttributes;
 import fll.web.AuthenticationContext;
 import fll.web.BaseFLLServlet;
@@ -63,6 +64,8 @@ import net.mtu.eggplant.xml.XMLUtils;
 public class SubjectiveByJudge extends BaseFLLServlet {
   private static final org.apache.logging.log4j.Logger LOGGER = org.apache.logging.log4j.LogManager.getLogger();
 
+  private static final String HIGHLIGHT_BACKGROUND_COLOR = "yellow";
+
   @Override
   protected void processRequest(HttpServletRequest request,
                                 HttpServletResponse response,
@@ -82,7 +85,7 @@ public class SubjectiveByJudge extends BaseFLLServlet {
 
     try (Connection connection = datasource.getConnection()) {
       ScoreStandardization.computeSummarizedScoresIfNeeded(connection, challengeDescription,
-                                                          tournamentData.getCurrentTournament());
+                                                           tournamentData.getCurrentTournament());
 
       final Tournament tournament = tournamentData.getCurrentTournament();
 
@@ -121,9 +124,11 @@ public class SubjectiveByJudge extends BaseFLLServlet {
     rootElement.appendChild(layoutMasterSet);
 
     final String pageMasterName = "simple";
-    final Element pageMaster = FOPUtils.createSimplePageMaster(document, pageMasterName);
+    final Element pageMaster = FOPUtils.createSimplePageMaster(document, pageMasterName,
+                                                               FOPUtils.PAGE_LANDSCAPE_LETTER_SIZE,
+                                                               new Margins(0.2, 0.2, 0.2, 0.2), 0,
+                                                               FOPUtils.STANDARD_FOOTER_HEIGHT);
     layoutMasterSet.appendChild(pageMaster);
-    pageMaster.setAttribute("reference-orientation", "90");
 
     final Element pageSequence = FOPUtils.createPageSequence(document, pageMasterName);
     rootElement.appendChild(pageSequence);
@@ -140,8 +145,8 @@ public class SubjectiveByJudge extends BaseFLLServlet {
     for (final String judgingGroup : Queries.getJudgingStations(connection, tournament.getTournamentID())) {
       final Element ele = generateAwardGroupReport(connection, document, challengeDescription, tournament, teams,
                                                    judgingGroup);
+      ele.setAttribute("keep-together.within-page", "always");
       documentBody.appendChild(ele);
-      ele.setAttribute("page-break-after", "always");
     }
 
     return document;
@@ -172,8 +177,8 @@ public class SubjectiveByJudge extends BaseFLLServlet {
             final String judge = castNonNull(rs.getString("judge"));
             final String station = castNonNull(rs.getString("station"));
 
-            final int numScores = ComputeJudgeSummary.getNumScoresEntered(connection, judge, category.getName(), station,
-                                                                      tournament.getTournamentID());
+            final int numScores = ComputeJudgeSummary.getNumScoresEntered(connection, judge, category.getName(),
+                                                                          station, tournament.getTournamentID());
             if (numScores > 0) {
               judges.add(judge);
             }
@@ -344,14 +349,22 @@ public class SubjectiveByJudge extends BaseFLLServlet {
                                                .getOrDefault(category, Collections.emptyMap())//
                                                .get(judge);
 
+          final boolean highlight;
           final String text;
           if (null != rankData) {
+            highlight = rankData.rank == 1;
             text = String.format("%d - %s", rankData.rank, rankData.rawScore);
           } else {
+            highlight = false;
             text = Utilities.NON_BREAKING_SPACE_STRING;
           }
-          teamRow.appendChild(FOPUtils.addBorders(FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, text),
-                                                  FOPUtils.STANDARD_BORDER_WIDTH));
+          final Element dataCell = FOPUtils.addBorders(FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER,
+                                                                                text),
+                                                       FOPUtils.STANDARD_BORDER_WIDTH);
+          if (highlight) {
+            dataCell.setAttribute("background-color", HIGHLIGHT_BACKGROUND_COLOR);
+          }
+          teamRow.appendChild(dataCell);
         } // foreach judge
       } // foreach category
     } // foreach team